+++ README.md
@@ -10,9 +10,8 @@
     - [x]  게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)
 - [x]  플레이어별로 히트/스탠드를 입력받는다.
     - [x]  pobi는 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)
-- [ ] 플레이어별로 배팅 금액을 입력받는다.
-    - [ ] 배팅 금액은 음수가 아니어야 한다.
-    - [ ] 배팅 금액은 짝수여야 한다.
+- [x] 플레이어별로 배팅 금액을 입력받는다.
+    - [x] 배팅 금액은 음수가 아니어야 한다.
 
 ## 출력
 
@@ -49,7 +48,7 @@
   pobi: 승 
   jason: 패
   ```
-- [ ] 최종 수익을 출력한다.
+- [x] 최종 수익을 출력한다.
 
 ``` markdown
 딜러: 10000
@@ -64,15 +63,15 @@ jason: -20000
 - [x]  Ace 1/11 처리
 - [x]  버스트 판정
 - [x]  히트/스탠드 처리
-- [ ] 플레이어 버스트 시 더 이상 카드를 뽑을 수 없다.
-- [ ]  승패 판정
-- [ ] 플레이어와 딜러 모두 블랙잭이면 무승부 처리한다.
-- [ ] 딜러가 21을 초과하면 버스트가 아닌 모든 플레이어는 승리 처리한다.
-- [ ] 배팅 기능
-    - [ ] 일반 승리 시 배팅 금액의 1.0배를 받는다.
-    - [ ] 블랙잭 승리 시 배팅 금액의 1.5배를 받는다.
-    - [ ] 일반 패배 시 배팅 금액을 모두 잃는다.
-    - [ ] 버스트 시 경우 배팅 금액을 모두 잃는다.
+- [x] 플레이어 버스트 시 더 이상 카드를 뽑을 수 없다.
+- [x] 승패 판정
+- [x] 플레이어와 딜러 모두 블랙잭이면 무승부 처리한다.
+- [x] 딜러가 21을 초과하면 버스트가 아닌 모든 플레이어는 승리 처리한다.
+- [x] 배팅 기능
+    - [x] 일반 승리 시 배팅 금액의 1.0배를 받는다.
+    - [x] 블랙잭 승리 시 배팅 금액의 1.5배를 받는다.
+    - [x] 일반 패배 시 배팅 금액을 모두 잃는다.
+    - [x] 버스트 시 경우 배팅 금액을 모두 잃는다.
 
 ## 유효성 검사
 
@@ -84,9 +83,3 @@ jason: -20000
 ### 히트/스탠드
 
 - [x]  y 또는 n이 아닌 값을 입력한 경우
-
-### 배팅 금액
-
-- [ ] 짝수가 아닌 경우
-- [ ] 양수가 아닌 경우
-- [ ] 숫자가 아닌 값을 입력한 경우
