+++ src/main/java/domain/dto/RoundResult.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.dto;
 
+import domain.BetMoney;
 import domain.participant.Player;
 
 public record RoundResult(
