+++ src/main/java/domain/dto/TotalResult.java
@@ -1,8 +1,10 @@
-package domain;
+package domain.dto;
+
+import domain.BetMoney;
 
 import java.util.List;
 
-public record Profits(
+public record TotalResult(
         List<RoundResult> results,
         BetMoney dealerProfit
 ) {
